refactor: review page
