+++ .vscode/app-snippet.code-snippets
@@ -1,0 +1,129 @@
+{
+	"Next Client Component": {
+		"prefix": "cfc",
+		"body": [
+		"\"use client\"",
+		"import { cn } from '@/utils/cn'",
+		"import  { HTMLAttributes } from 'react'",
+		"",
+		"interface Props extends HTMLAttributes<HTMLDivElement> {",
+		" $1",
+		"}",
+		"",
+		"export const ${TM_FILENAME_BASE/([a-z]*)-*([a-z]*)/${1:/capitalize}${2:/capitalize}/g} = ({ className }: Props) => {",
+		"  return <div className={cn(className)}>${TM_FILENAME_BASE/([a-z]*)-*([a-z]*)/${1:/capitalize}${2:/capitalize}/g}</div>",
+		"}",
+		"",
+		],
+		"description": "Typescript React Function Component"
+	},
+	"Next Component": {
+		"prefix": "fc",
+		"body": [
+		"import { cn } from '@/shared/lib/utils'",
+		"import  { HTMLAttributes } from 'react'",
+		"",
+		"interface Props extends HTMLAttributes<HTMLDivElement> {",
+		" $1",
+		"}",
+		"",
+		"export const ${TM_FILENAME_BASE/([a-z]*)-*([a-z]*)/${1:/capitalize}${2:/capitalize}/g} = ({ className }: Props) => {",
+		"  return <div className={cn(className)}>${TM_FILENAME_BASE/([a-z]*)-*([a-z]*)/${1:/capitalize}${2:/capitalize}/g}</div>",
+		"}",
+		"",
+		],
+		"description": "Typescript React Function Component"
+	},
+	
+	"Next Layout": {
+		"prefix": "layout",
+		"body": [
+		"import { FunctionComponent, PropsWithChildren } from 'react'",
+		"import type { Metadata } from 'next'",
+	
+		"export const metadata: Metadata = {",
+		" title: 'Create Next App',",
+		" description: 'Generated by create next app',",
+		"}",
+	
+		"interface $1${TM_FILENAME_BASE/([a-z]*)-*([a-z]*)/${1:/capitalize}${2:/capitalize}/g}Props extends PropsWithChildren {",
+		"",
+		"}",
+		"",
+		"const $1${TM_FILENAME_BASE/([a-z]*)-*([a-z]*)/${1:/capitalize}${2:/capitalize}/g}: FunctionComponent<$1${TM_FILENAME_BASE/([a-z]*)-*([a-z]*)/${1:/capitalize}${2:/capitalize}/g}Props> = ({children}) => {",
+		"  return <main>{children}</main>",
+		"}",
+		"",
+		"export default $1${TM_FILENAME_BASE/([a-z]*)-*([a-z]*)/${1:/capitalize}${2:/capitalize}/g}"
+		],
+		"description": "Nextjs Layout File"
+	},
+	
+	"Next Error": {
+		"prefix": "error-page",
+		"body":[
+		"'use client' // Error components must be Client Components"
+	
+		"import { useEffect } from 'react'"
+		
+		"export default function Error({"
+		"  error,"
+		"  reset,"
+		"}: {"
+		"  error: Error & { digest?: string }"
+		"  reset: () => void"
+		"}) {"
+		"  useEffect(() => {"
+		"    // Log the error to an error reporting service"
+		"    console.error(error)"
+		"  }, [error])"
+		""
+		"  return ("
+		"    <div>"
+		"      <h2>Something went wrong!</h2>"
+		"      <button"
+		"        onClick={"
+		"          // Attempt to recover by trying to re-render the segment"
+		"          () => reset()"
+		"        }"
+		"      >"
+		"        Try again"
+		"      </button>"
+		"    </div>"
+		"  )"
+		"}"
+	],
+		"description": "Nextjs Error Page"
+	},
+	
+	"Next Page": {
+		"prefix": "page",
+		"body": [
+		
+		"const ${TM_FILENAME_BASE/([a-z]*)-*([a-z]*)/${1:/capitalize}${2:/capitalize}/g} = async () => {",
+		"  return <main>$1${TM_FILENAME_BASE/([a-z]*)-*([a-z]*)/${1:/capitalize}${2:/capitalize}/g}</main>",
+		"}",
+		"",
+		"export default ${TM_FILENAME_BASE/([a-z]*)-*([a-z]*)/${1:/capitalize}${2:/capitalize}/g}"
+		],
+		"description": "Nextjs Page"
+	},
+	"Next Detail Page": {
+		"prefix": "dpage",
+		"body": [
+		"interface Props {",
+		"    params: {",
+		"        id: string",
+		"    }",
+		"}",
+		"",
+		"const $1Page = async ({params: {id}}: Props) => {",
+		"  return <>$2</>",
+		"}",
+		"",
+		"export default $1Page"
+		],
+		"description": "Nextjs Detail Page"
+	}
+}
+
