+++ src/shared/components/icon/svg-components.tsx
@@ -1,0 +1,304 @@
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+function generateUUID() {
+  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
+    const r = (Math.random() * 16) | 0,
+      v = c == 'x' ? r : (r & 0x3) | 0x8
+    return v.toString(16)
+  })
+}
+
+export function Ranking({ ...props }) {
+  return (
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path
+        d="M8.87891 6.79139C8.87891 6.16447 9.38712 5.65625 10.014 5.65625H13.987C14.6139 5.65625 15.1221 6.16447 15.1221 6.79139V22.1157H8.87891V6.79139Z"
+        fill="#95B0F8"
+      />
+      <path
+        d="M1.5 11.8988C1.5 11.2719 2.00822 10.7637 2.63514 10.7637H6.60811C7.23503 10.7637 7.74324 11.2719 7.74324 11.8988V22.115H1.5V11.8988Z"
+        fill="#7095F8"
+      />
+      <path
+        d="M16.2568 14.1693C16.2568 13.5424 16.7651 13.0342 17.392 13.0342H21.3649C21.9919 13.0342 22.5001 13.5424 22.5001 14.1693V22.1153H16.2568V14.1693Z"
+        fill="#95B0F8"
+      />
+      <circle cx="16.2266" cy="5.625" r="3.375" fill="#FF9100" />
+      <path
+        d="M14.877 5.45547L15.8895 6.63672L17.577 4.94922"
+        stroke="#F5F7F9"
+        stroke-width="0.675"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+      />
+    </svg>
+  )
+}
+
+export function Stars({ ...props }) {
+  return (
+    <svg
+      width="19"
+      height="27"
+      viewBox="0 0 19 27"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path
+        d="M13.3348 15.4661C13.6806 14.8551 14.4681 14.8551 14.8139 15.4661L16.1516 17.8298C16.2079 17.9292 16.2793 18.0165 16.3625 18.0875L18.4148 19.8405C18.8816 20.2392 18.8816 21.0338 18.4148 21.4324L16.3625 23.1854C16.2793 23.2565 16.2079 23.3438 16.1516 23.4432L14.8139 25.8069C14.4681 26.4179 13.6806 26.4179 13.3348 25.8069L11.9971 23.4432C11.9408 23.3438 11.8694 23.2565 11.7862 23.1854L9.73384 21.4324C9.2671 21.0338 9.2671 20.2392 9.73384 19.8405L11.7862 18.0875C11.8694 18.0165 11.9408 17.9292 11.9971 17.8298L13.3348 15.4661Z"
+        fill="#FFAB40"
+      />
+      <path
+        d="M5.53144 0.611017C6.01556 -0.203673 7.118 -0.203672 7.60212 0.611017L9.47495 3.76262C9.55371 3.89516 9.65366 4.01152 9.77014 4.10627L12.6435 6.44361C13.2969 6.97515 13.2969 8.03462 12.6435 8.56617L9.77014 10.9035C9.65366 10.9983 9.55371 11.1146 9.47495 11.2472L7.60212 14.3988C7.118 15.2135 6.01556 15.2135 5.53144 14.3988L3.65861 11.2472C3.57985 11.1146 3.47989 10.9983 3.36341 10.9035L0.490077 8.56617C-0.163358 8.03463 -0.163359 6.97516 0.490076 6.44361L3.36341 4.10627C3.47989 4.01152 3.57985 3.89516 3.65861 3.76262L5.53144 0.611017Z"
+        fill="#FFAB40"
+      />
+    </svg>
+  )
+}
+
+export function Check({ ...props }) {
+  return (
+    <svg
+      width="16"
+      height="12"
+      viewBox="0 0 16 12"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path
+        d="M2 5.42857L6.2 10L14 2"
+        stroke="#63CF75"
+        strokeWidth="3"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  )
+}
+
+export function QuizArchive({ ...props }) {
+  return (
+    <svg
+      width="33"
+      height="27"
+      viewBox="0 0 33 27"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <rect x="4.49512" y="2.22754" width="20.2744" height="22.2524" fill="#BCCFFF" />
+      <rect x="4.49512" y="22.502" width="17.3074" height="3.95597" rx="1.97799" fill="#BCCFFF" />
+      <path
+        d="M0.538086 3.71156C0.538086 3.98466 0.75948 4.20605 1.03258 4.20605H10.428C10.9742 4.20605 11.417 3.76327 11.417 3.21706V2.96981C11.417 1.46774 10.1994 0.25008 8.69728 0.25008H2.51607C1.42366 0.25008 0.538086 1.13565 0.538086 2.22807V3.71156Z"
+        fill="#95B0F8"
+      />
+      <path
+        d="M6.54793 26.459C7.68166 26.459 8.60073 25.4929 8.60073 24.3012L8.60073 22.9975C8.60073 22.7244 8.82213 22.503 9.09523 22.503L26.1412 22.503C26.4143 22.503 26.6357 22.7244 26.6357 22.9975L26.6357 24.3012C26.6979 25.0204 26.4117 26.459 24.7695 26.459L6.54793 26.459Z"
+        fill="#95B0F8"
+      />
+      <path
+        d="M2.45598 0.25C3.58208 0.25 4.49497 1.21608 4.49497 2.4078L4.49497 4.20597L24.7684 4.20597L24.7684 2.4078C24.8302 1.68854 24.546 0.250001 22.9148 0.250001L2.45598 0.25Z"
+        fill="#BCCFFF"
+      />
+      <path
+        d="M8.20312 4.7002H21.06"
+        stroke="#D7E2FF"
+        stroke-width="0.988994"
+        stroke-linecap="round"
+      />
+      <path
+        d="M8.45117 7.17285H21.3081"
+        stroke="#D7E2FF"
+        stroke-width="0.988994"
+        stroke-linecap="round"
+      />
+      <path
+        d="M8.45117 9.64551H21.3081"
+        stroke="#D7E2FF"
+        stroke-width="0.988994"
+        stroke-linecap="round"
+      />
+      <path
+        d="M8.45117 12.1182H14.3851"
+        stroke="#D7E2FF"
+        stroke-width="0.988994"
+        stroke-linecap="round"
+      />
+      <path
+        d="M14.9054 19.4173C14.5028 19.5785 14.1051 19.175 14.2721 18.7747L15.5874 15.6223C15.722 15.2996 16.1419 15.2182 16.3874 15.4673L18.2432 17.3504C18.4887 17.5994 18.4012 18.0181 18.0766 18.148L14.9054 19.4173Z"
+        fill="#FFECD0"
+      />
+      <path
+        d="M14.8994 19.4235C14.4958 19.5867 14.096 19.1811 14.2651 18.7799L14.7131 17.7171L15.9686 18.991L14.8994 19.4235Z"
+        fill="#4B4F54"
+      />
+      <path
+        d="M26.7294 4.31773C26.922 4.12823 27.2317 4.13076 27.4212 4.32338L29.5465 6.48359C29.736 6.67621 29.7334 6.98598 29.5408 7.17549L18.6344 17.9052C18.4388 18.0976 18.1232 18.0916 17.9351 17.8918L16.9851 16.8827L15.8232 15.7469C15.6273 15.5554 15.6268 15.2404 15.8221 15.0483L26.7294 4.31773Z"
+        fill="#7095F8"
+      />
+      <rect
+        x="29.0225"
+        y="2.09766"
+        width="4.00887"
+        height="2.18104"
+        rx="0.489259"
+        transform="rotate(45.4679 29.0225 2.09766)"
+        fill="#FF9100"
+      />
+      <rect
+        x="26.9326"
+        y="3.83105"
+        width="4.42675"
+        height="0.706322"
+        rx="0.353161"
+        transform="rotate(45.4679 26.9326 3.83105)"
+        fill="#D2D6DB"
+      />
+      <rect
+        x="27.4199"
+        y="3.32617"
+        width="4.45971"
+        height="0.726188"
+        rx="0.363094"
+        transform="rotate(45.4679 27.4199 3.32617)"
+        fill="#D2D6DB"
+      />
+    </svg>
+  )
+}
+
+export function SavePick({ ...props }) {
+  return (
+    <svg
+      width="31"
+      height="31"
+      viewBox="0 0 31 31"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <rect x="5.8125" y="3.875" width="21.3125" height="27.125" rx="2.90625" fill="#FF9100" />
+      <rect x="5.8125" y="3.875" width="20.3438" height="26.1562" rx="2.90625" fill="#F5F7F9" />
+      <rect x="4.84375" y="2.90625" width="19.8996" height="25.3891" rx="2.05858" fill="#FFD180" />
+      <path
+        d="M4.84375 4.96483C4.84375 3.8279 5.76541 2.90625 6.90233 2.90625H8.27471V28.2954H6.90233C5.76541 28.2954 4.84375 27.3737 4.84375 26.2368V4.96483Z"
+        fill="#FFAB40"
+      />
+      <rect x="10.334" y="7.02344" width="11.6653" height="4.11716" rx="0.686193" fill="#FFE1AC" />
+      <path
+        d="M29.1184 11.8334L25.4797 9.51696L21.8409 11.8334C21.6 11.9857 21.2832 11.8173 21.2832 11.5341V1.67805C21.2832 1.13836 21.7272 0.700195 22.2741 0.700195H28.6852C29.2321 0.700195 29.6761 1.13836 29.6761 1.67805V11.5315C29.6761 11.8147 29.3594 11.9857 29.1184 11.8307V11.8334Z"
+        fill="#7095F8"
+      />
+    </svg>
+  )
+}
+
+export function ChevronRight({ ...props }) {
+  return (
+    <svg
+      width="6"
+      height="10"
+      viewBox="0 0 6 10"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path
+        d="M1 9L5 5L1 1"
+        stroke="#D2D6DB"
+        stroke-width="2"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+      />
+    </svg>
+  )
+}
+
+export function NoAnalysis({ ...props }) {
+  return (
+    <svg
+      width="85"
+      height="87"
+      viewBox="0 0 85 87"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <rect x="12.9297" y="14.6533" width="47.7044" height="52.3585" fill="#EAECEF" />
+      <rect x="12.9297" y="62.3574" width="40.7233" height="9.30818" rx="4.65409" fill="#EAECEF" />
+      <path
+        d="M3.61914 18.1451C3.61914 18.7877 4.14007 19.3086 4.78266 19.3086H26.8896C28.1748 19.3086 29.2166 18.2667 29.2166 16.9815V16.3998C29.2166 12.8655 26.3515 10.0004 22.8173 10.0004H8.27323C5.70285 10.0004 3.61914 12.0841 3.61914 14.6545V18.1451Z"
+        fill="#D2D6DB"
+      />
+      <path
+        d="M17.7598 71.668C20.4274 71.668 22.59 69.3948 22.59 66.5908L22.59 63.5233C22.59 62.8807 23.1109 62.3598 23.7535 62.3598L63.8616 62.3598C64.5042 62.3598 65.0251 62.8807 65.0251 63.5233L65.0251 66.5908C65.1715 68.2832 64.4982 71.668 60.6341 71.668L17.7598 71.668Z"
+        fill="#D2D6DB"
+      />
+      <path
+        d="M8.1316 10C10.7813 10 12.9292 12.2731 12.9292 15.0772L12.9292 19.3082L60.6315 19.3082L60.6315 15.0772C60.7768 13.3848 60.1081 10 56.27 10L8.1316 10Z"
+        fill="#EAECEF"
+      />
+      <path
+        d="M21.6543 20.4707H51.9059"
+        stroke="#F5F7F9"
+        strokeWidth="2.32704"
+        strokeLinecap="round"
+      />
+      <path
+        d="M22.2383 26.2891H52.4899"
+        stroke="#F5F7F9"
+        strokeWidth="2.32704"
+        strokeLinecap="round"
+      />
+      <path
+        d="M22.2383 32.1074H52.4899"
+        stroke="#F5F7F9"
+        strokeWidth="2.32704"
+        strokeLinecap="round"
+      />
+      <path
+        d="M22.2383 37.9248H36.2005"
+        stroke="#F5F7F9"
+        strokeWidth="2.32704"
+        strokeLinecap="round"
+      />
+      <rect x="63" y="52.002" width="4" height="4" rx="0.833333" fill="#797D81" />
+      <circle cx="65" cy="69.916" r="15" fill="#EAECEF" />
+      <circle cx="65" cy="69.916" r="13.9583" stroke="#A2A6AB" strokeWidth="2.08333" />
+      <path
+        d="M77.4123 56.1144C76.6312 55.3334 75.3649 55.3334 74.5838 56.1144L72.7071 57.9911L76.3153 61.5993L78.192 59.7226C78.973 58.9415 78.973 57.6752 78.192 56.8941L77.4123 56.1144Z"
+        fill="#A2A6AB"
+      />
+      <g filter="url(#filter0_f_2873_6441)">
+        <path
+          d="M65 59.916C63.3555 59.916 61.7364 60.3216 60.286 61.0968C58.8357 61.872 57.5989 62.993 56.6853 64.3603C55.7717 65.7277 55.2093 67.2993 55.0482 68.9358C54.887 70.5724 55.1319 72.2235 55.7612 73.7428C56.3905 75.2622 57.3848 76.6029 58.6561 77.6461C59.9273 78.6894 61.4362 79.403 63.0491 79.7239C64.662 80.0447 66.3291 79.9628 67.9028 79.4854C69.4765 79.008 70.9082 78.1499 72.0711 76.9871L65 69.916V59.916Z"
+          fill="#D2D6DB"
+        />
+      </g>
+      <rect x="60.832" y="49.084" width="8.33333" height="4.16667" rx="0.833333" fill="#A2A6AB" />
+      <defs>
+        <filter
+          id="filter0_f_2873_6441"
+          x="53"
+          y="57.916"
+          width="21.0713"
+          height="24"
+          filterUnits="userSpaceOnUse"
+          colorInterpolationFilters="sRGB"
+        >
+          <feFlood floodOpacity="0" result="BackgroundImageFix" />
+          <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape" />
+          <feGaussianBlur stdDeviation="1" result="effect1_foregroundBlur_2873_6441" />
+        </filter>
+      </defs>
+    </svg>
+  )
+}
